Move each function into its own file.
This is so that one could require JUST the function they need instead of all other functions associated to a utility.
