+++ index.js
@@ -1,5 +1,4 @@
-var api = {};
-api.url = require('./lib/url');
-
-module.exports = api;
+module.exports = {
+  url: require('./lib/url')
+};
