+++ lib/url/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  cleanString: require('./cleanString'),
+  getPath: require('./getPath'),
+  isFullUrl: require('./isFullUrl'),
+  replaceBaseWith: require('./replaceBaseWith')
+};
